+++ lib/src/resultset-table/ResultsetTable.tsx
@@ -1,9 +1,8 @@
 import React, { useState, useMemo, useEffect } from "react";
 import styled, { ThemeProvider } from "styled-components";
 import { spaces } from "../common/variables";
-import DxcActionsCell from "../table/ActionsCell";
 import DxcPaginator from "../paginator/Paginator";
-import DxcTable from "../table/Table";
+import DxcTable, { DxcActionsCell } from "../table/Table";
 import useTheme from "../useTheme";
 import ResultsetTablePropsType, { Column } from "./types";
 import icons from "./Icons";
