+++ lib/src/table/Table.tsx
@@ -3,9 +3,65 @@ import styled, { ThemeProvider } from "styled-components";
 import { spaces } from "../common/variables";
 import { getMargin } from "../common/utils";
 import useTheme from "../useTheme";
-import TablePropsType from "./types";
-import CoreTokens from "../common/coreTokens";
-import DxcActionsCell from "./ActionsCell";
+import TablePropsType, { ActionCellsPropsType, DeepPartial } from "./types";
+import DropdownTheme from "./DropdownTheme";
+import { AdvancedTheme } from "../common/variables";
+import { DxcDropdown, DxcFlex, HalstackProvider } from "../main";
+import DxcActionIcon from "../action-icon/ActionIcon";
+
+const moreVertIcon = (
+  <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 -960 960 960" width="24" fill="currentColor">
+    <path d="M480-160q-33 0-56.5-23.5T400-240q0-33 23.5-56.5T480-320q33 0 56.5 23.5T560-240q0 33-23.5 56.5T480-160Zm0-240q-33 0-56.5-23.5T400-480q0-33 23.5-56.5T480-560q33 0 56.5 23.5T560-480q0 33-23.5 56.5T480-400Zm0-240q-33 0-56.5-23.5T400-720q0-33 23.5-56.5T480-800q33 0 56.5 23.5T560-720q0 33-23.5 56.5T480-640Z" />
+  </svg>
+);
+
+const overwriteTheme = (theme: DeepPartial<AdvancedTheme>) => {
+  const newTheme = DropdownTheme;
+  newTheme.dropdown.buttonBackgroundColor = theme.table.actionBackgroundColor;
+  newTheme.dropdown.hoverButtonBackgroundColor = theme.table.hoverActionBackgroundColor;
+  newTheme.dropdown.activeButtonBackgroundColor = theme.table.activeActionBackgroundColor;
+  newTheme.dropdown.buttonIconColor = theme.table.actionIconColor;
+  newTheme.dropdown.disabledColor = theme.table.disabledActionIconColor;
+  newTheme.dropdown.disabledButtonBackgroundColor = theme.table.disabledActionBackgroundColor;
+  return newTheme;
+};
+
+export const DxcActionsCell = ({ actions }: ActionCellsPropsType): JSX.Element => {
+  const actionIcons = actions.filter((action) => !action.options);
+  const actionDropdown = actions.find((action) => action.options);
+  const maxNumberOfActions = actionDropdown ? 2 : 3;
+  const colorsTheme = useTheme();
+
+  return (
+    <DxcFlex gap={"0.5rem"} alignItems="center">
+      {actionIcons.map(
+        (action, index) =>
+          index < maxNumberOfActions && (
+            <DxcActionIcon
+              icon={action.icon}
+              title={action.title}
+              onClick={action.onClick}
+              disabled={action.disabled ?? false}
+              tabIndex={action.tabIndex ?? 0}
+              key={`action-${index}`}
+            />
+          )
+      )}
+      {actionDropdown && (
+        <HalstackProvider advancedTheme={overwriteTheme(colorsTheme)} key={`provider-dropdown`}>
+          <DxcDropdown
+            options={actionDropdown.options}
+            onSelectOption={actionDropdown.onClick}
+            disabled={actionDropdown.disabled}
+            icon={moreVertIcon}
+            tabIndex={actionDropdown.tabIndex}
+            caretHidden
+          ></DxcDropdown>
+        </HalstackProvider>
+      )}
+    </DxcFlex>
+  );
+};
 
 const DxcTable = ({ children, margin, mode = "default" }: TablePropsType): JSX.Element => {
   const colorsTheme = useTheme();
@@ -91,17 +147,21 @@ const DxcTableContent = styled.table<{ mode: TablePropsType["mode"] }>`
   }
   & th:first-child {
     border-top-left-radius: ${(props) => props.theme.headerBorderRadius};
-    padding-left: ${(props) => (props.mode === "default" ? props.theme.firstChildPaddingLeft : props.theme.firstChildPaddingLeftReduced)};
+    padding-left: ${(props) =>
+      props.mode === "default" ? props.theme.firstChildPaddingLeft : props.theme.firstChildPaddingLeftReduced};
   }
   & th:last-child {
     border-top-right-radius: ${(props) => props.theme.headerBorderRadius};
-    padding-right: ${(props) => (props.mode === "default" ? props.theme.lastChildPaddingRight : props.theme.lastChildPaddingRightReduced)};
+    padding-right: ${(props) =>
+      props.mode === "default" ? props.theme.lastChildPaddingRight : props.theme.lastChildPaddingRightReduced};
   }
   & td:first-child {
-    padding-left: ${(props) => (props.mode === "default" ? props.theme.firstChildPaddingLeft : props.theme.firstChildPaddingLeftReduced)};
+    padding-left: ${(props) =>
+      props.mode === "default" ? props.theme.firstChildPaddingLeft : props.theme.firstChildPaddingLeftReduced};
   }
   & td:last-child {
-    padding-right: ${(props) => (props.mode === "default" ? props.theme.lastChildPaddingRight : props.theme.lastChildPaddingRightReduced)};
+    padding-right: ${(props) =>
+      props.mode === "default" ? props.theme.lastChildPaddingRight : props.theme.lastChildPaddingRightReduced};
   }
 `;
